midi_handling/requirements.txt
mido 
python-rtmidi
pygame
pydub
pyfluidsynth